requirements.txt
numpy==1.18.5
pandas==1.1.5
Pillow==8.0.1
librosa==0.8.0
tensorflow==2.3.1
scikit-learn==0.23.2
matplotlib~=3.3.3
tqdm~=4.54.1
seaborn~=0.11.0